Add text support function
Added a support function to render text to a surface. The advantage of
this function over other functions in pygame is that the blit location
is expressed in relative, not absolute terms.

Also moved the LevelSplash class to the same file as the Level class,
since it's a small, lightweight class that doesn't need to be in its
own file.