Refactor ship engines into separate class
Refactored the ship engines into their own classes. MainEngine and
RotationEngine both inherit from Engine, and all Engines contain
Flames. There are a few advantages to factoring engines out, but the
chief difference from the player's perspective is that engine thrust
and fuel consumption now ramps up and down in the same manner that the
flame animations do. This makes the gameplay slightly more realistic,
and better aligns the physics with the visuals.