Fix bug that allows for course cutting
Fixed an embarrassing bug in the program that allowed for course
cutting. The FinishBox did not have any safeguards to ensure that all
gates had been completed before it could be activated. A savvy player
could fly straight into the FinishBox and immediately finish the race.
This bug has been fixed, so all gates must be activated before the
FinishBox will do anything.