Use Vector2 class for coordinates
The game uses a lot of Cartesian coordinate pairs, which up until now
I had been representing with pairs of variables such as position_x and
position_y, and passing 2-tuples around as necessary. Pygame includes a
Vector2 class that makes all this manipulation a lot easier, so even
though it's an experimental class, I'm switching over.

In the process, I've also simplified the internal semantics of angles
to be consistent with pygame, so all angles are now represented as
degrees counterclockwise from (1, 0), instead of radians
counterclockwise from (0, 1).