Implement new waypoint type: Proxy
Implemented a new type of waypoint I'm calling a Proxy for now (short
for Proximity). Its behavior differs from a Gate in a couple of ways:
first, Proxies do not need to be activated in a certain order. As long
as they all are activated before entering the FinishBox, no problem!
Second, Proxies don't have any orientation -- as long as the player's
ship gets within a certain radius of them, they will activate.

Proxy support is limited at this point. They do not show up in the HUD,
and the FinishBox doesn't enforce proxy completion yet, but the kernel
of basic behavior is there.