Create GateSet and ProxySet classes
There was a lot of messy detail for the sets of gates and proxies
within the Course class, so I refactored the class to contain separate
GateSet and ProxySet classes that handle all of the details. Every
Course will have both a GateSet and a ProxySet, regardless of whether
it has any Gates (or Proxies). This avoids a lot of annoying
conditionals within the Course class itself.