Add Fuel and FuelTank classes
I refactored the fuel elements of the Ship class into separate Fuel and
FuelTank classes. A Ship contains a FuelTank, which contains a quantity
of Fuel. Besides cleaning up the code, this move makes a few planned
enhancements down the road easier to implement:
  * Multiple types of fuel are available, each with their own density
    and exit velocity.
  * The mass of the fuel tank is independent of the mass of the ship
    (but is still considered when calculating the ship's acceleration).
    This can allow the player to choose from one of several sizes of
    fuel tank.
  * The same ship can have multiple tanks, each with its own type of
    fuel. For example, a ship can use a bipropellant for its main
    engine and a monopropellant for its directional engines.

The one immediate change to gameplay is that the efficiency of the
engine is now a function of how full the fuel tank is. The engine is at
maximum efficiency with a full tank, and efficiency declines very
slowly as fuel depletes.