Make Ship an abstract base class
My plan for a while has been to have multiple types of ships available
in the game, and it seems the most sensible way to implement this is by
making a Ship abstract base class and a separate subclass for each type
of Ship.

The first subclass of Ship is Pegasus, which is the design of ship that
I've been testing with so far. The intention is for the Pegasus to be
the first/simplest ship class, and the the exception of some minor
parameter tweaking, I'm not planning on adding any major new features
to the Pegasus class.