Dynamically rescale minimap
One problem I uncovered while watching others try to play this game
this weekend is that the player often got fairly off course and then
found it virtually impossible to navigate back to the course. This is
largely because once the player flies sufficiently off-course, there
are no visual indications whatsoever as to the direction of the course.

I fixed this issue by modifying the minimap to rescale itself to
automatically keep the Ship (and the entire course) in-frame at all
times.

In the process, I also refactored the bounding_box property of the
Level class to return a Rect instead of a 4-tuple.