Overhaul HUD design
Overhauled the design of the game HUD to fill out the available space
more, make it easier to add additional fields, and improve readability.

Some fields on the HUD (notably Split and Proxies) are dummies that
have not been implemented yet; others (like Fuel) still have some major
overhauling left, and the whole HUD still has myriad tweaks of sizes,
positions, and colors ahead, but this new layout is much closer to the
final version.