Construct ship from YAML file
Created a function ship_from_dict() that constructs a Ship object from
a YAML file with the ship's configuration settings. I also modified the
level constructor to accept a dictionary of ship parameters and a
dictionary of course parameters, and added another argument to
ZeroGee.py for a YAML file with the ship configuration.