Implement player profiles
Implemented player profiles, which for now only track best times and
splits on each course. The program now automatically loads a default
profile, and overwrites the best time and splits for a course if a new
record is set.

In the process, caught and fixed a bug in the HUD where negative split
times would be displayed incorrectly.