Create Sprite abstract class for rendered objects
Added a new abstract class Sprite from which objects that are rendered
on the screen will inherit from, and a LevelSprite class which is
specifically adapted for the needs of Sprites that are part of primary
gameplay.

The goals of the class are twofold: first, to consolidate and
deduplicate a lot of boilerplate rendering code across the project.
Second, there are foreseeable contexts in which we may wish to
instantiate some objects without wanting to draw them (for example,
validating a sequence of keypresses will yield a certain time on a
given course for record-keeping, or training an AI racer). By creating
(for example), a Gate class with just the gameplay logic and a
GateSprite class with the rendering logic, this becomes much easier.

Right now, the only classes that have transitioned to use Sprite as a
parent are the classes in waypoint, but more will follow.